Django_Backend/DjangoMedicalStoreManagementSystem/requirements.txt
asgiref==3.8.1
django==5.1.1
django-cors-headers==4.5.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
markdown==3.7
pyjwt==2.9.0
python-dotenv==1.0.1
sqlparse==0.5.1
typing-extensions==4.12.2
tzdata==2024.1
